feature/conjured test_standard_items
